requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
html5lib==1.1
idna==2.10
python-decouple==3.4
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.3
webencodings==0.5.1
